+++ Source/HelixToolkit.SharpDX.Shared/Core/VolumeRenderCore.cs
@@ -18,10 +18,12 @@ namespace HelixToolkit.UWP
 {
     namespace Core
     {
+        using Components;
         using Model;
         using Render;
         using Shaders;
         using System.Runtime.CompilerServices;
+        using System.Runtime.InteropServices;
         using Utilities;
 
         public sealed class VolumeRenderCore : RenderCore
@@ -61,12 +63,28 @@ namespace HelixToolkit.UWP
                 };
             }
 
+            [StructLayout(LayoutKind.Sequential, Pack = 4)]
+            struct ModelMatrices{
+                public Matrix ModelMatrix;
+                public Matrix ModelMatrixInv;
+                public void Update(ref Matrix modelMatrix)
+                {
+                    if(ModelMatrix != modelMatrix)
+                    {
+                        ModelMatrix = modelMatrix;
+                        ModelMatrixInv = modelMatrix.Inverted();
+                    }
+                }
+            }
+
             private VolumeCubeBufferModel buffer;
-        
+
             private ShaderPass cubeBackPass;
             private ShaderPass volumePass;
+            private ShaderPass meshFrontPass;
             private int backTexSlot;
-
+            private readonly ConstantBufferComponent modelCB;
+            private ModelMatrices modelMatrices;
             private MaterialVariable materialVariables = EmptyMaterialVariable.EmptyVariable;
             /// <summary>
             /// Used to wrap all material resources
@@ -90,9 +108,11 @@ namespace HelixToolkit.UWP
                 }
             }
 
-            public VolumeRenderCore() 
+            public VolumeRenderCore()
                 : base(RenderType.Opaque)
             {
+                modelCB = Collect(new ConstantBufferComponent(new ConstantBufferDescription(DefaultBufferNames.VolumeModelCB, 
+                    VolumeParamsStruct.SizeInBytes)));
             }
 
             protected override bool OnAttach(IRenderTechnique technique)
@@ -101,29 +121,59 @@ namespace HelixToolkit.UWP
                 buffer.Geometry = BoxMesh;
                 buffer.Topology = PrimitiveTopology.TriangleList;
                 cubeBackPass = technique[DefaultPassNames.Backface];
+                meshFrontPass = technique[DefaultPassNames.Positions];
+                modelCB.Attach(technique);
                 return true;
             }
 
             protected override void OnDetach()
             {
                 buffer = null;
+                modelCB.Detach();
                 base.OnDetach();
             }
 
             public override void Render(RenderContext context, DeviceContextProxy deviceContext)
             {
-                if (!materialVariables.UpdateMaterialStruct(deviceContext, ref ModelMatrix, Matrix.SizeInBytes))
-                {
-                    return;
-                }
-                int slot = 0;
-                buffer.AttachBuffers(deviceContext, ref slot, EffectTechnique.EffectsManager);
-                cubeBackPass.BindShader(deviceContext);
-                cubeBackPass.BindStates(deviceContext, StateType.All);
                 using (var back = context.GetOffScreenRT(OffScreenTextureSize.Full, global::SharpDX.DXGI.Format.R16G16B16A16_Float))
                 {
-                    BindTarget(null, back, deviceContext, (int)context.ActualWidth, (int)context.ActualHeight);
-                    deviceContext.DrawIndexed(buffer.IndexBuffer.ElementCount, 0, 0);
+                    int slot = 0;
+                    using (var depth = context.GetOffScreenDS(OffScreenTextureSize.Full, global::SharpDX.DXGI.Format.D32_Float_S8X24_UInt))
+                    {
+                        deviceContext.ClearDepthStencilView(depth, DepthStencilClearFlags.Depth | DepthStencilClearFlags.Stencil, 1, 1);
+                        deviceContext.ClearRenderTargetView(back, new Color4(0, 0, 0, 0));
+                        BindTarget(depth, back, deviceContext, (int)context.ActualWidth, (int)context.ActualHeight);
+                        #region Render box back face and set stencil buffer to 0
+                        modelMatrices.Update(ref ModelMatrix);
+                        if (!materialVariables.UpdateMaterialStruct(deviceContext, ref modelMatrices, Matrix.SizeInBytes * 2))
+                        {
+                            return;
+                        }
+                        buffer.AttachBuffers(deviceContext, ref slot, EffectTechnique.EffectsManager);
+                        cubeBackPass.BindShader(deviceContext);
+                        cubeBackPass.BindStates(deviceContext, StateType.All);
+                        deviceContext.DrawIndexed(buffer.IndexBuffer.ElementCount, 0, 0);
+                        #endregion
+
+                        #region Render all mesh Positions onto off-screen texture region with stencil = 0 only
+                        if (context.RenderHost.PerFrameOpaqueNodesInFrustum.Count > 0)
+                        {
+                            for (int i = 0; i < context.RenderHost.PerFrameOpaqueNodesInFrustum.Count; ++i)
+                            {
+                                var mesh = context.RenderHost.PerFrameOpaqueNodesInFrustum[i];
+                                var meshPass = mesh.EffectTechnique[DefaultPassNames.Positions];
+                                if (meshPass.IsNULL) { continue; }
+                                meshPass.BindShader(deviceContext);
+                                meshPass.BindStates(deviceContext, StateType.BlendState);
+                                // Set special depth stencil state to only render into region with stencil region is 0
+                                meshFrontPass.BindStates(deviceContext, StateType.DepthStencilState);
+                                mesh.RenderCustom(context, deviceContext);
+                            }
+                        }
+                        #endregion
+                    }
+
+                    #region Render box back face again and do actual volume sampling
                     context.RenderHost.SetDefaultRenderTargets(false);
                     var pass = materialVariables.GetPass(RenderType.Opaque, context);
                     if (pass != volumePass)
@@ -131,11 +181,14 @@ namespace HelixToolkit.UWP
                         volumePass = pass;
                         backTexSlot = volumePass.PixelShader.ShaderResourceViewMapping.TryGetBindSlot(DefaultBufferNames.VolumeBack);
                     }
+                    slot = 0;
+                    buffer.AttachBuffers(deviceContext, ref slot, EffectTechnique.EffectsManager);
                     materialVariables.BindMaterialResources(context, deviceContext, pass);
                     volumePass.PixelShader.BindTexture(deviceContext, backTexSlot, back);
                     volumePass.BindShader(deviceContext);
                     volumePass.BindStates(deviceContext, StateType.All);
                     deviceContext.DrawIndexed(buffer.IndexBuffer.ElementCount, 0, 0);
+                    #endregion
                     volumePass.PixelShader.BindTexture(deviceContext, backTexSlot, null);
                 }
             }