+++ Source/HelixToolkit.SharpDX.Shared/Model/Layouts.cs
@@ -404,6 +404,8 @@ namespace HelixToolkit.UWP
         public const string HasTextureStr = "bHasTexture"; //bool
         public const string TextureScaleStr = "pTextureScale";//float;
         public const string AlphaThresholdStr = "pAlphaThreshold"; // float; 
+        public const string EnableBlendingStr = "pEnableBlending"; //bool
+        public const string BlendingFactorStr = "pBlendingFactor"; //float
     }
 
     /// <summary>
@@ -562,6 +564,7 @@ namespace HelixToolkit.UWP
     public struct VolumeParamsStruct
     {
         public const string World = "mWorld"; //Separated from the struct in material
+        public const string WorldInv = "mWorldInv"; // Inverse of world matrix
         public const string Color = "pColor"; // Vector4
         public const string StepSize = "stepSize"; // Vector3
         public const string MaxIterations = "maxIterations"; // int or uint
@@ -569,9 +572,9 @@ namespace HelixToolkit.UWP
         public const string IsoValue = "isoValue"; // float
         public const string BaseSampleDistance = "baseSampleDist"; //float
         public const string ActualSampleDistance = "actualSampleDist"; //float
-        public const string ScaleFactor = "scaleFactor"; //Vector4
         public const string IterationOffset = "iterationOffset"; // int or uint
-        public const int SizeInBytes = 4 * (4 * 4 + 4 + 4 + 4 + 4);
+        public const string EnablePlaneAlignment = "enablePlaneAlignment"; // bool
+        public const int SizeInBytes = 4 * (4 * 4 + 4 * 4 + 4 + 4 + 4);
     }
 
     /// <summary>