+++ Source/HelixToolkit.SharpDX.Shared/Model/Material/VolumeTextureMaterial.cs
@@ -99,6 +99,8 @@ namespace HelixToolkit.UWP
             /// The transfer map.
             /// </value>
             Color4[] TransferMap { set; get; }
+
+            bool EnablePlaneAlignment { set; get; }
         }
 
         /// <summary>
@@ -134,7 +136,7 @@ namespace HelixToolkit.UWP
                 get { return sampleDistance; }
             }
 
-            private int maxIterations = int.MaxValue;
+            private int maxIterations = 512;
             /// <summary>
             /// Gets or sets the iteration. Usually set to VolumeDepth.
             /// </summary>
@@ -194,6 +196,12 @@ namespace HelixToolkit.UWP
                 get { return transferMap; }
             }
 
+            private bool enablePlaneAlignment = true;
+            public bool EnablePlaneAlignment
+            {
+                set { Set(ref enablePlaneAlignment, value); }
+                get { return enablePlaneAlignment; }
+            }
             protected virtual string DefaultPassName { get; } = DefaultPassNames.Default;
 
             public override MaterialVariable CreateMaterialVariables(IEffectsManager manager, IRenderTechnique technique)